FILE: Moved pty-specific code to ptyconnection.cpp.
